更新代码

+++ src/common/ERC20.sol
@@ -4,41 +4,33 @@ pragma solidity ^0.8.26;
 import "./IERC20.sol";
 
 contract ERC20 is IERC20 {
-
     mapping(address => uint256) public override balanceOf;
     mapping(address => mapping(address => uint256)) public override allowance;
-    uint256 public override totalSupply;   // 代币总供给
+    uint256 public override totalSupply; // 代币总供给
 
-    string public name;   // 名称
-    string public symbol;  // 代号
+    string public name; // 名称
+    string public symbol; // 代号
     uint8 public decimals = 18; // 小数位数
 
-
-    constructor(string memory name_, string memory symbol_){
+    constructor(string memory name_, string memory symbol_) {
         name = name_;
         symbol = symbol_;
     }
 
-    function transfer(address recipient, uint amount) public override returns (bool) {
+    function transfer(address recipient, uint256 amount) public override returns (bool) {
         balanceOf[msg.sender] -= amount;
         balanceOf[recipient] += amount;
         emit Transfer(msg.sender, recipient, amount);
         return true;
     }
 
-
-    function approve(address spender, uint amount) public override returns (bool) {
+    function approve(address spender, uint256 amount) public override returns (bool) {
         allowance[msg.sender][spender] = amount;
         emit Approval(msg.sender, spender, amount);
         return true;
     }
 
-
-    function transferFrom(
-        address sender,
-        address recipient,
-        uint amount
-    ) public override returns (bool) {
+    function transferFrom(address sender, address recipient, uint256 amount) public override returns (bool) {
         allowance[sender][msg.sender] -= amount;
         balanceOf[sender] -= amount;
         balanceOf[recipient] += amount;
@@ -46,27 +38,23 @@ contract ERC20 is IERC20 {
         return true;
     }
 
-
-    function _mint(address to,uint amount) private {
+    function _mint(address to, uint256 amount) private {
         balanceOf[to] += amount;
         totalSupply += amount;
         emit Transfer(address(0), to, amount);
     }
 
-
-    function _burn(address from,uint amount) private {
+    function _burn(address from, uint256 amount) private {
         balanceOf[from] -= amount;
         totalSupply -= amount;
         emit Transfer(from, address(0), amount);
     }
 
-
-    function mint(address to,uint amount) external{
+    function mint(address to, uint256 amount) external {
         _mint(to, amount);
     }
 
-
-    function burn(address from,uint amount) external{
+    function burn(address from, uint256 amount) external {
         _burn(from, amount);
     }
 }
